attempt at creating hook
