+++ frontend/app/src/hooks/use-pending-invites.ts
@@ -1,0 +1,37 @@
+import { useQuery } from '@tanstack/react-query';
+import api from '@/lib/api';
+import { cloudApi } from '@/lib/api/api';
+import useCloudApiMeta from '@/pages/auth/hooks/use-cloud-api-meta';
+
+export const usePendingInvites = () => {
+  const cloudMeta = useCloudApiMeta();
+
+  const query = useQuery({
+    queryKey: ['pending-invites'],
+    queryFn: async () => {
+      const [tenantInvites, orgInvites] = await Promise.allSettled([
+        api.userListTenantInvites(),
+        cloudMeta?.data
+          ? cloudApi.userListOrganizationInvites()
+          : Promise.resolve({ data: { rows: [] } }),
+      ]);
+
+      const tenantCount =
+        tenantInvites.status === 'fulfilled'
+          ? tenantInvites.value.data.rows?.length || 0
+          : 0;
+      const orgCount =
+        orgInvites.status === 'fulfilled'
+          ? orgInvites.value.data.rows?.length || 0
+          : 0;
+
+      return tenantCount + orgCount;
+    },
+    refetchInterval: 30000, // Refetch every 30 seconds
+    enabled: true,
+  });
+
+  return {
+    pendingInvitesQuery: query,
+  };
+};
